pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=2.1.0",
    "django>=5.2.9",
    "django-crispy-forms>=2.5",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "groq>=0.18.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "weasyprint>=67.0",
    "whitenoise>=6.11.0",
]
